move test to  a proper class
