+++ tests/Install-Scripts.Test/GivenThatIWantToInstallDotnetFromAScript.cs
@@ -595,6 +595,24 @@ namespace Microsoft.DotNet.InstallationScript.Tests
             commandResult.Should().NotHaveStdOutContainingIgnoreCase(feedCredential);
         }
 
+        [Theory]
+        [InlineData("7.0.1", Quality.Daily)]
+        [InlineData("6.0.1", Quality.Signed)]
+        public void WhenBothVersionAndQualityWereSpecified(string version, Quality quality)
+        {
+            string feedCredentials = Guid.NewGuid().ToString();
+            var args = new[] { "-dryrun", "-version", version, "-quality", quality.ToString(), "-feedCredential", feedCredentials };
+
+            var commandResult = CreateInstallCommand(args)
+                            .CaptureStdOut()
+                            .CaptureStdErr()
+                            .Execute();
+
+            commandResult.Should().Fail();
+            commandResult.Should().HaveStdErrContaining("Quality and Version options are not allowed to be specified simultaneously.");
+            commandResult.Should().NotHaveStdOutContaining("Installation finished");
+        }
+
         private static IEnumerable<string> GetInstallScriptArgs(
             string? channel, 
             string? runtime,
